feat(memory): move generated glossary to long memory
Render generated glossary notes as Tier 2 long memory with an indexed
term and alias description.

Exclude stale generated glossary short notes from AMD inlining during
the one-pass migration.

Regenerate instruction artifacts and update docs and tests for the new
glossary memory contract.

SASE_PLAN=[202608/glossary_long_memory.md][1]
SASE_TYPE=stitch
SASE_AGENT=[bbugyi200.athena.ys][2]

[1]: https://github.com/sase-org/sase--plans/blob/main/202608/glossary_long_memory.md
[2]: https://github.com/sase-org/sase--agents/blob/main/families/bbugyi200.athena.ys.md

+++ tests/main/test_xprompt_handler.py
@@ -46,7 +46,7 @@ def test_xprompt_list_includes_kind_and_insertion(
         "memory/glossary": Workflow(
             name="memory/glossary",
             source_path="/tmp/sase/memory/glossary.md",
-            memory_type="short",
+            memory_type="long",
             steps=[WorkflowStep(name="prompt", prompt_part="Glossary body")],
         ),
         "multi": _xprompt_swarm_workflow("multi"),
@@ -75,7 +75,7 @@ def test_xprompt_list_includes_kind_and_insertion(
     assert rows["memory/glossary"]["kind"] == "memory"
     assert rows["memory/glossary"]["prefix"] == "#"
     assert rows["memory/glossary"]["insertion"] == "#memory/glossary"
-    assert rows["memory/glossary"]["memory_type"] == "short"
+    assert rows["memory/glossary"]["memory_type"] == "long"
     assert rows["memory/glossary"]["source"] == "/tmp/sase/memory/glossary.md"
 
     assert rows["multi"]["type"] == "xprompt"
